LayoutTests/fast/forms/auto-fill-button/mouse-down-input-mouse-release-auto-fill-button-expected.txt

This test checks that pressing the mouse in an password field and releasing the mouse on the AutoFill button dispatches a DOM click event at the password field and not at the AutoFill button. This test can only be run in the test tool.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



PASS event.type is "click"
PASS event.target is document.getElementById("password")
PASS successfullyParsed is true

TEST COMPLETE
